% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Find geopolitical place from location string}
\usage{
where(x)
}
\arguments{
\item{x}{Input should be a character vector containing location strings or
a data frame with a 'location' column.}
}
\value{
Returns a data.table with the original data (character strings
  returned in the 'location' column) joined with geopolitical columns
}
\description{
Converts a location string into a geopolitical place–based matches to cities
states, or countries
}
\details{
This isn't exact. It's very much based on trying to make reasonable
  guesses.
}
